+++ modules/payments-component/src/paymets/payment-methods/stripe/validators/only-number.validator.ts
@@ -1,0 +1,12 @@
+import {AbstractControl} from "@angular/forms";
+
+export function onlyNumberValidator(control: AbstractControl): {[key: string]: any} {
+  let NUMBER_REGEXP = /^[0-9]+$/;
+  if (control.value && !NUMBER_REGEXP.test(control.value)) {
+    return {
+      validNumber: {
+        valid: false
+      }
+    };
+  }
+}